requirements.txt
requests==2.27.1
python-dotenv==0.20.0
python-telegram-bot==20.0a2